pyproject.toml
[tool.poetry]
name = "melqui-system"
version = "0.1.0"
description = ""
authors = ["Danilo Nascimento <daconnas.dcn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
python-dotenv = "^1.0.1"
schema = "^0.7.7"
psycopg2-binary = "^2.9.10"
psycopg = {extras = ["pool"], version = "^3.2.3"}
django-import-export = "^4.2.0"
django-crispy-forms = "^2.3"
crispy-bootstrap4 = "^2024.10"
django-admin-rangefilter = "^0.13.2"
openpyxl = "^3.1.5"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"
blue = "^0.9.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
